+++ Covid/common.h
@@ -37,6 +37,6 @@ constexpr f32 t_05_100 = 1.982f;
 
 constexpr int NUM_EXPERIMENTS = 101;
 
-constexpr float SCREEN_WIDTH = 1024;
-constexpr float SCREEN_HEIGHT = 1024;
+constexpr i32 SCREEN_WIDTH = 1024;
+constexpr i32 SCREEN_HEIGHT = 1024;
 //constexpr int NUM_EXPERIMENTS = 101;
